+++ src/client.ts
@@ -35,6 +35,7 @@ import {
   initializeOpenOrdersIx,
   closeOpenOrdersIx,
   placeOrderIx,
+  placeOrderV2Ix,
   depositIx,
   withdrawIx,
   cancelOrderIx,
@@ -449,7 +450,6 @@ export class Client {
    * @param price           the native price of the order (6 d.p as integer)
    * @param size            the quantity of the order (3 d.p)
    * @param side            the side of the order. bid / ask
-   * @param orderType       the type of the order. limit / ioc / post-only
    * @param clientOrderId   optional: client order id (non 0 value)
    * NOTE: If duplicate client order ids are used, after a cancel order,
    * to cancel the second order with the same client order id,
@@ -458,6 +458,74 @@ export class Client {
    * (Depending on the order in which the order was cancelled).
    */
   public async placeOrder(
+    market: PublicKey,
+    price: number,
+    size: number,
+    side: Side,
+    clientOrderId = 0
+  ): Promise<TransactionSignature> {
+    let tx = new Transaction();
+    let marketIndex = Exchange.markets.getMarketIndex(market);
+
+    let openOrdersPda = null;
+    if (this._openOrdersAccounts[marketIndex].equals(PublicKey.default)) {
+      console.log(
+        `User doesn't have open orders account. Initialising for market ${market.toString()}.`
+      );
+      let [initIx, _openOrdersPda] = await initializeOpenOrdersIx(
+        market,
+        this.publicKey,
+        this.marginAccountAddress
+      );
+      openOrdersPda = _openOrdersPda;
+      tx.add(initIx);
+    } else {
+      openOrdersPda = this._openOrdersAccounts[marketIndex];
+    }
+
+    let orderIx = placeOrderIx(
+      marketIndex,
+      price,
+      size,
+      side,
+      clientOrderId,
+      this.marginAccountAddress,
+      this.publicKey,
+      openOrdersPda,
+      this._whitelistTradingFeesAddress
+    );
+
+    tx.add(orderIx);
+
+    let txId: TransactionSignature;
+    try {
+      this._openOrdersAccounts[marketIndex] = openOrdersPda;
+      txId = await utils.processTransaction(this._provider, tx);
+    } catch (e) {
+      // If we were initializing open orders in the same tx.
+      if (tx.instructions.length > 1) {
+        this._openOrdersAccounts[marketIndex] = PublicKey.default;
+      }
+      throw e;
+    }
+    return txId;
+  }
+
+  /**
+   * Places an order on a zeta market.
+   * @param market          the address of the serum market
+   * @param price           the native price of the order (6 d.p as integer)
+   * @param size            the quantity of the order (3 d.p)
+   * @param side            the side of the order. bid / ask
+   * @param orderType       the type of the order. limit / ioc / post-only
+   * @param clientOrderId   optional: client order id (non 0 value)
+   * NOTE: If duplicate client order ids are used, after a cancel order,
+   * to cancel the second order with the same client order id,
+   * you may need to crank the corresponding event queue to flush that order id
+   * from the user open orders account before cancelling the second order.
+   * (Depending on the order in which the order was cancelled).
+   */
+  public async placeOrderV2(
     market: PublicKey,
     price: number,
     size: number,
@@ -484,7 +552,7 @@ export class Client {
       openOrdersPda = this._openOrdersAccounts[marketIndex];
     }
 
-    let orderIx = placeOrderIx(
+    let orderIx = placeOrderV2Ix(
       marketIndex,
       price,
       size,
@@ -572,9 +640,57 @@ export class Client {
    * @param newOrderprice  the native price of the order (6 d.p) as integer
    * @param newOrderSize   the quantity of the order (3 d.p) as integer
    * @param newOrderside   the side of the order. bid / ask
-   * @param newOrderType   the type of the order, limit / ioc / post-only
    */
   public async cancelAndPlaceOrder(
+    market: PublicKey,
+    orderId: anchor.BN,
+    cancelSide: Side,
+    newOrderPrice: number,
+    newOrderSize: number,
+    newOrderSide: Side,
+    clientOrderId = 0
+  ): Promise<TransactionSignature> {
+    let tx = new Transaction();
+    let marketIndex = Exchange.markets.getMarketIndex(market);
+    let ixs = [];
+    ixs.push(
+      cancelOrderIx(
+        marketIndex,
+        this.publicKey,
+        this._marginAccountAddress,
+        this._openOrdersAccounts[marketIndex],
+        orderId,
+        cancelSide
+      )
+    );
+    ixs.push(
+      placeOrderIx(
+        marketIndex,
+        newOrderPrice,
+        newOrderSize,
+        newOrderSide,
+        clientOrderId,
+        this.marginAccountAddress,
+        this.publicKey,
+        this._openOrdersAccounts[marketIndex],
+        this._whitelistTradingFeesAddress
+      )
+    );
+    ixs.forEach((ix) => tx.add(ix));
+    return await utils.processTransaction(this._provider, tx);
+  }
+
+  /**
+   * Cancels a user order by orderId and atomically places an order
+   * @param market     the market address of the order to be cancelled.
+   * @param orderId    the order id of the order.
+   * @param cancelSide       the side of the order. bid / ask.
+   * @param newOrderprice  the native price of the order (6 d.p) as integer
+   * @param newOrderSize   the quantity of the order (3 d.p) as integer
+   * @param newOrderside   the side of the order. bid / ask
+   * @param newOrderType   the type of the order, limit / ioc / post-only
+   */
+  public async cancelAndPlaceOrderV2(
     market: PublicKey,
     orderId: anchor.BN,
     cancelSide: Side,
@@ -598,7 +714,7 @@ export class Client {
       )
     );
     ixs.push(
-      placeOrderIx(
+      placeOrderV2Ix(
         marketIndex,
         newOrderPrice,
         newOrderSize,
@@ -622,10 +738,56 @@ export class Client {
    * @param newOrderprice           the native price of the order (6 d.p) as integer
    * @param newOrderSize            the quantity of the order (3 d.p) as integer
    * @param newOrderSide            the side of the order. bid / ask
-   * @param newOrderType            the type of the order, limit / ioc / post-only
    * @param newOrderClientOrderId   the client order id for the new order
    */
   public async cancelAndPlaceOrderByClientOrderId(
+    market: PublicKey,
+    cancelClientOrderId: number,
+    newOrderPrice: number,
+    newOrderSize: number,
+    newOrderSide: Side,
+    newOrderClientOrderId: number
+  ): Promise<TransactionSignature> {
+    let tx = new Transaction();
+    let marketIndex = Exchange.markets.getMarketIndex(market);
+    let ixs = [];
+    ixs.push(
+      cancelOrderByClientOrderIdIx(
+        marketIndex,
+        this.publicKey,
+        this._marginAccountAddress,
+        this._openOrdersAccounts[marketIndex],
+        new anchor.BN(cancelClientOrderId)
+      )
+    );
+    ixs.push(
+      placeOrderIx(
+        marketIndex,
+        newOrderPrice,
+        newOrderSize,
+        newOrderSide,
+        newOrderClientOrderId,
+        this.marginAccountAddress,
+        this.publicKey,
+        this._openOrdersAccounts[marketIndex],
+        this._whitelistTradingFeesAddress
+      )
+    );
+    ixs.forEach((ix) => tx.add(ix));
+    return await utils.processTransaction(this._provider, tx);
+  }
+
+  /**
+   * Cancels a user order by client order id and atomically places an order by new client order id.
+   * @param market                  the market address of the order to be cancelled and new order.
+   * @param cancelClientOrderId     the client order id of the order to be cancelled.
+   * @param newOrderprice           the native price of the order (6 d.p) as integer
+   * @param newOrderSize            the quantity of the order (3 d.p) as integer
+   * @param newOrderSide            the side of the order. bid / ask
+   * @param newOrderType            the type of the order, limit / ioc / post-only
+   * @param newOrderClientOrderId   the client order id for the new order
+   */
+  public async cancelAndPlaceOrderByClientOrderIdV2(
     market: PublicKey,
     cancelClientOrderId: number,
     newOrderPrice: number,
@@ -647,7 +809,7 @@ export class Client {
       )
     );
     ixs.push(
-      placeOrderIx(
+      placeOrderV2Ix(
         marketIndex,
         newOrderPrice,
         newOrderSize,